Changing the probability from Double to Int

+++ app/io/perfix/model/ValueProbability.scala
@@ -2,7 +2,7 @@ package io.perfix.model
 
 import play.api.libs.json.{Format, Json}
 import PerfixQuestionAnswer._
-case class ValueProbability(value: Any, probability: Double)
+case class ValueProbability(value: Any, probability: Int)
 
 object ValueProbability {
   implicit val ValueProbability: Format[ValueProbability] = Json.format[ValueProbability]
